Remove ClientsideMessageEvent
