+++ common/src/main/java/com/wynntils/mc/event/ChatPacketReceivedEvent.java
@@ -4,6 +4,7 @@
  */
 package com.wynntils.mc.event;
 
+import com.wynntils.core.text.StyledText;
 import net.minecraft.network.chat.Component;
 import net.neoforged.bus.api.Event;
 import net.neoforged.bus.api.ICancellableEvent;
@@ -11,9 +12,21 @@ import net.neoforged.bus.api.ICancellableEvent;
 public abstract class ChatPacketReceivedEvent extends Event implements ICancellableEvent {
     private Component message;
     private boolean messageChanged;
+    private final StyledText originalStyledText;
+    private StyledText styledText;
 
     protected ChatPacketReceivedEvent(Component message) {
         this.message = message;
+        this.originalStyledText = StyledText.fromComponent(message);
+        this.styledText = originalStyledText;
+    }
+
+    public StyledText getOriginalStyledText() {
+        return originalStyledText;
+    }
+
+    public StyledText getStyledText() {
+        return styledText;
     }
 
     public Component getMessage() {
@@ -23,6 +36,7 @@ public abstract class ChatPacketReceivedEvent extends Event implements ICancella
     public void setMessage(Component message) {
         this.message = message;
         this.messageChanged = true;
+        this.styledText = StyledText.fromComponent(message);;
     }
 
     public boolean isMessageChanged() {
